LayoutTests/imported/w3c/web-platform-tests/close-watcher/iframes/dialog-cross-origin-ynn.html

<!doctype html>
<meta name="timeout" content="long">
<link rel="author" href="mailto:lwarlow@igalia.com" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="../resources/helpers.js"></script>
<script src="/common/get-host-info.sub.js"></script>

<body>
  <iframe id="iframe1"></iframe>
  <iframe id="iframe2"></iframe>

  <script>
    const host_info = get_host_info();

    iframe1.src =
        `${host_info.HTTP_REMOTE_ORIGIN}` +
        `/close-watcher/iframes/resources/dialog-prevents-close.html`;
    iframe2.src =
        `${host_info.HTTP_REMOTE_ORIGIN}` +
        `/close-watcher/iframes/resources/dialog-prevents-close.html`;

    function awaitEvent(el, type, signal) {
      return new Promise((resolve) =>
        el.addEventListener(type, resolve, { once: true, signal }),
      );
    }

    async function iframeDialogIsOpen(iframe, signal) {
      const reply = awaitEvent(window, "message", signal);
      iframe.contentWindow.postMessage("dialog_open", "*");
      const {data} = (await reply);
      if (data.error) throw new Error(data.error);
      return data.open;
    }

    promise_test(async (t) => {
      // The order the iframes load is inconsistent across engines,
      // and not the subject of this test, so just await independent of ordering.
      await Promise.all([
          awaitEvent(iframe1, "load", t.get_signal()),
          awaitEvent(iframe2, "load", t.get_signal()),
      ]);

      await test_driver.bless();

      assert_true(await iframeDialogIsOpen(iframe1, t.get_signal()), "Dialog 1 is open");
      assert_true(await iframeDialogIsOpen(iframe2, t.get_signal()), "Dialog 2 is open");

      await test_driver.send_keys(iframe1, "\uE00C");

      assert_false(await iframeDialogIsOpen(iframe1, t.get_signal()), "Dialog 1 is now closed");
      assert_true(await iframeDialogIsOpen(iframe2, t.get_signal()), "Dialog 2 is still open");

      await test_driver.send_keys(iframe2, "\uE00C");

      assert_false(await iframeDialogIsOpen(iframe1, t.get_signal()), "Dialog 1 is still closed");
      assert_false(await iframeDialogIsOpen(iframe2, t.get_signal()), "Dialog 2 is now closed");
    });
  </script>
</body>
